Disable DNS Client Cache service
Our DNS server seems to be ignored after a while when the DNS Client
Cache is enabled. Disabling it makes it so DNS resolution will always go
through the server and caching is left up for each client. In our case,
we TTL set to 0 that means no caching.

It's not clear if disabling that service has great implications for
other things but it seems like many other people have successfully
disabled it without major consequences.

Found another story from the Greenhouse team where DNS Client Cache was
re-enabled in #107746116. This was after it was initially disabled
in #91881972 for the same reasons we were experiencing in testing.

[#144881929](https://www.pivotaltracker.com/story/show/144881929)

Signed-off-by: Luan Santos <lsantos@pivotal.io>