Revert "Use RecursionAvailable flag"
This is currently failing in test-unit-windows due to recursion supposedly not being available.

    https://main.bosh-ci.cf-app.com/teams/main/pipelines/bosh-dns-release/jobs/test-unit-windows/builds/290

This may be due to the Windows worker not being configured for DNS, or the Windows forwarder
not knowing how to automatically load the system recursors. Essentially, the reverted commit
added conditional behavior based on the length of the configured recursors list and, if
recursors were not configured, the test results are dependent on the environment.

This still needs to be investigated, fixed, and re-committed.

This reverts commit 0e297ba72f309405c5070edec2e6250eb6ff0dd7.

[#151540552](https://www.pivotaltracker.com/story/show/151540552)