Set ServerPriorityTimeLimit to 0 on Windows
Instead of disabling the Windows DNS cache service (which is not
possible in Windows Server 2016), or setting the MaxCacheTtl to 0, just
set the ServerPriorityTimeLimit to ensure that BOSH DNS is always tried
first.

We already set MaxNegativeCacheTtl, so records should always be looked
up in BOSH DNS for internal domains. This will also allow external
domains and consul records to be cached, without the need for the
enable_os_dns_caching flag.