Switch from MAC detection to rely on bind address
Filtering virtual interfaces by MAC address proved incorrect because for
a Virtual Machine a Hyper-V Network interface, for example, is
recognized as a physical interface.

Instead, since we will only ever be updating the primary interface where
the server is bound, find the adapter based on that address and use the
DNS servers listed on that interface.

[finishes #157235521](https://www.pivotaltracker.com/story/show/157235521)

Signed-off-by: Chris Dutra <cdutra@pivotal.io>