Report group specific healthiness in health server
This change adds the concept of group specific healthiness to the
response returned by the health server.

Currently, the health server will parse group metadata from the jobs
directory at the path `.bosh/links.json`. This file will contain a JSON
array in the format: `[{"group":"id"}...]`.

The health server will also detect the presence of the health executable
script when iterating over the jobs defined in the jobs directory.

If a group is defined in a job that contains a health executable, then
it's state will be reflected by the return value of the health
executable. It will no longer be tied to the overall VM health. This
allows for more granular definitions of health and is a prerequisite to
link specific healthy addressing.

If a group is defined in a job that does not contain a health
executable, then it's state will match the overall VM health.

[#160701107](https://www.pivotaltracker.com/story/show/160701107)
[#160701135](https://www.pivotaltracker.com/story/show/160701135)

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>
Co-authored-by: Mike Xu <mxu@pivotal.io>
Co-authored-by: Danny Berger <dberger@pivotal.io>