Remove fs_repo_performance_test.go
Testing performance by measurign a very specific procedure on unit tests
has a tendendcy to be very brittle. It is very difficult to predict how
fast a machine running the tests will be. This causes us to have to
adjust this number based on the slowest computer, which defeats the
purpuse of having it in the first place.

Decision for this was made as a group here:
https://cloudfoundry.slack.com/archives/C094QRLEQ/p1541637322227300

Co-authored-by: Luan Santos <lsantos@pivotal.io>