Use xenial stemcell for recursor/http-dns-server
((base_stemcell)) used to fill in both windows and linux manifests.
Since there is not a need to test the recursor test dependency on
windows, it is better to simply pick a stemcell for it to run with.
This was causing failures in trusty, however. The test-acceptance linux
concourse task would previously only upload a single stemcell for the
DNS job and use it for the DNS manifest, the test-recursor, and the
http-dns-server deployments. Since only one stemcell was being uploaded,
this could result in failure when the stemcell was hard-coded to xenial
for test-recursor/http-dns-server deployments and the DNS deployment was
using trusty.  For now, it seems easier to explicitly upload xenial in
test-acceptance in addition to the base-stemcell (which is a duplicate
upload in the case of testing xenial, but should not be a problem for
testing).

[#160999831](https://www.pivotaltracker.com/story/show/160999831)