This test suite relies on no DNS recursors
This test was failing due to test-pollution since we were not removing
the test recursor from the cloud-config.  Since the test recursor does
not respond with NXDOMAIN when getting an upcheck.bosh-dns. query we
need to delete all DNS servers from the cloud-config to preserve the
previous behavior.  The purpose of the test is to ensure that the
bosh-dns server itself does not get prepended to resolv.conf, and since
the name resolution is not getting handled within the VM in any case we
can be sure this is working properly.

[#160999831](https://www.pivotaltracker.com/story/show/160999831)