Implement serial recursor selection
This commit adds a new property which controls the recursor selection.
The two possible values are "smart" and "serial". The default is
"smart".

When configured to be "serial", bosh-dns will use recursors to resolve
upstream addresses in the order that the recursors are listed in
`resolv.conf` or in the explicit bosh-dns recursor configuration.

This commit also contains the following:

 - Refactor recursor_test.go
   - Reduce redundant tests
   - Add tests for different recursor_selection

[#165801868]

Co-authored-by: Joshua Aresty <joshua.aresty@emc.com>