Make handler configuration example correct
In the first example given, the domain is set to `local.internal.`
and the `url` is set to `endpoint.local`. In the second, the domain
is set to `local.internal2.` Given that `.local` is a common TLD,
the example is confusing because it implies that the DNS names should
be listed in reverse order (akin to Java packages). This PR corrects
the first case, and hopefully makes both examples more readable.