Use `spec.ip` as default listen address for bosh-dns-windows
In 1.18 the default listen address was changed to 127.0.0.1 so the dns
server would only be available locally on each VM, to match the default
behaviour on linux. When two non-loopback non-tunnel network interfaces
are present bosh-dns-windows failed to start due to how the
implementation of `windows_manager.go` tries to identify the "primary"
adapter by matching the IP address.

[#171637356](https://www.pivotaltracker.com/story/show/171637356)