Log failed recurser requests at Error level
Current version of bosh-dns requires DEBUG level for recursor errors.
It is not easy to reproduce recursor errors and DEBUG level is too noisy
as production configuration. That is why it will be better to log
recursor failures at ERROR level. This change will log all recursor
errors at ERROR level beside the "NXDOMAIN" errror, which I think is
easy for debugging and could be too noise.

fixises #60