Makes QNAME handling case-insensitive
This commit changes bosh-dns's config file ingestion and query handling
to ensure that QNAME comparisons are done in a case-insensitive manner.

While the underlying DNS helper libraries that we were relying on all
correctly handled DNS operations in a case-insensitive manner, our
home-grown support code did not.

This commit also adds tests that scramble the case of QNAMEs and
input file contents to validate that the fixes actually fix things.

NOTE: The performance tests were not updated to scramble the case of
QNAMEs, but we expect that all of the other testing gives us a
sufficient amount of coverage.

[#178114087] Make bosh-dns consider DNS name lookups to be case-insensitive