Sets GODEBUG envvar for old CN-validation behavior
Customers may not yet have updated their TLS certs to have both SANs and
CNs. If we don't set this env var, then they will be in an inconvenient
world of partial-brokenness.

Testing reveals that executables that are altered by `chpst` do inherit
env vars that exist in the environment, so exporting the var should be
enough.

[#178114087] Make bosh-dns consider DNS name lookups to be case-insensitive

Signed-off-by: Joseph Palermo <jpalermo@vmware.com>