Fix CI deployment to work after nats fw agent feature
once we introduced the bosh agent nats outgoing firewall
feature a bosh deployed docker host could not be used any-
more by a cpi. This was because the docker container "vms"
were not able to talk to nats because the docker hosts nats
protection blocked outgoing traffic. this adds the pre start
script from os conf to remove the rule created on the docker
host so the container "vms" can talk to nats.