Update bosh-cli go dependency to package v7
We mistakenly did not correctly version the bosh-cli when we introduced go mod
to it. This prevented repos using it as a dependency, such as this one,
from being able to vendor in newer versions of the code.

We have updated the bosh-cli to correctly have a v7 package version. This
commit pulls in that new version as the imported package.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>