add option to test on jammy
in order to test on jammy we need to update asset manifests to use
base_stemcell variable. the current deployment commands executed within
ginkgo already set the value but the manifests are not aware of it.

start-bosh currently always will use jammy to deploy the director.
start-bosh is aware of STEMCELL_OS variable and can conditionally switch
the stemcell used to deploy the director.

add an additional job to run acceptance tests on jammy for proper
testing of the release for both currently supported stemcells in OSS.