Changes VM size used for Director and VMs
Recently, the default VM size changed to `t3.medium` which is not nearly
enough to run our stress tests.
<https://github.com/cloudfoundry/bosh-deployment/pull/435>

This commit uses the size that most closely matches the size that used
to be the defaults, but deploy on the latest-available machine types.

Hopefully all regions that we operate in have the type we've selected.
We shall see.

Signed-off-by: Kenneth Lakin <klakin@vmware.com>