Add golangci-lint and address issues
In almost all cases issues have been addressed by adding `//nolint:...`
directives.

There are now two lint scripts `scripts/lint` which defaults to checking
`./...` or the directory it is passes as a first argument. The second is
`scripts/lint-all` which iterates over a hard-coded list of directories
(see list below) derived from the `scripts/test-unit` contents, checking
each directory's set of pacakges (e.g. `./...` for that directory).

The `scripts/test-unit` script now runs `scripts/lint` for the
directories (these are the same set checked in `scripts/lint-all`):
- `src/bosh-dns/acceptance_tests/dns-acceptance-release/src/test-recursor`
- `src/bosh-dns`
- `src/debug`

as it builds utilities and/or runs tests.