Build the `test-recursor` binary via Ginkgo
This commit moves the responsibility for building the `test-recursor`
binary in
`src/bosh-dns/acceptance_tests/dns-acceptance-release/src/test-recursor`
from the bash and powershell scripts, into the Ginkgo test code.

This removes the need to use ENV variables to pass around the location
of the binary used by the integration-test suite.