Negative caching: we only negative-cache SOAs
We don't negative-cache the common records one would expect: address
records. Instead, we only negative-cache the SOA (start of authority)
record. This commit clarifies that by adding a specific test that shows
that when an upstream DNS server returns an NXDOMAIN ("not found"), BOSH
DNS will NOT negative-cache the result UNLESS the record type happens to
be an SOA.

This behavior is subtle, and not something we can easily change without
modifying CoreDNS internals.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>