Adds prints to better understand NXDOMAIN caching
This commit adds a whole bunch of print statements which may help me to
understand how bosh-dns is (or is not) caching NXDOMAIN Responses.

I wrote up a test (which will be in the next commit) to demonstrate that
bosh-dns DOES NOT cache NXDOMAIN responses... but discovered that they
appeared to be cached just fine.

Deeper printf-driven investigation ensued, and this is the result.