Adds NXDOMAIN caching test (and debug prints)
This commit augments the existing cache handler test to demonstrate that
probably-properly-formatted NXDOMAIN Responses do get cached, and
subsequent queries pull the result from the cache.

This commit also adds a bunch of debugging prints that should be
removed.
Additionally, the "when an answer is cached" set of tests DO NOT work as
expected... there may be some test setup issues that I don't have time
to get into at End Of Day.