Allows for caching of NXDOMAIN responses
Okay.

It turns out that if we do not consider NXDOMAIN responses as errors,
then they make their way into the CoreDNS caching system, and subsequent
queries for that record are served from the cache. The lifetime of the
cache entry is set to the TTL of the SOA record that's delivered along
with the NXDOMAIN response. (RFC2308, Section 5 says that NXDOMAIN
responses that do not carry an SOA record MUST NOT be cached.)
Experimentation suggests that we can tamper with this TTL value with no
ill effect, which gives us control over the lifetime of the cache entry.

Given that this is code that _we_ wrote, rather than library code we
depend on, plumbing configuration for this through to the correct spot
should be fairly simple.

But, why are things this way? I did some research and have a story:

In early 2020, the code was modified from "Treat only SERVFAIL (and
-presumably- query timeouts)  as a server error" to "Treat anything
other than NOERROR as a server error". [0][1]

Why? Because some customers configure BOSH-DNS with resolvers that are
incapable of reaching servers that are authoritative for every zone that
they might be queried for. The old behavior caused client applications
to receive NXDOMAIN responses for queries that COULD have been answered
by one or more of the currently-configured resolvers... but not by the
currently-selected-by-BOSH-DNS resolver. Marking NXDOMAIN responses as
"the server is busted" means that BOSH-DNS will flop over to the next
configured resolver and retry the query.

(One might argue that making the `handlers` property [2][3][4] (which
 allows associating a particular zone with a particular set of recursors)
 less cumbersome to use would have been the better way to deal with
 this, but here we are.)

Given that backwards-compatibility concerns require that we retain this
behavior, we'll need to make "Cache NXDOMAIN responses?" configurable,
and default to "No.". We SHOULD also make the NXDOMAIN response cache
time configurable... with the default being what CoreDNS is using to set
the TTL of the SOA record that comes along with the NXDOMAIN response.

Given that customers are known to sometimes have absurdly-broken
configurations that they cannot reasonably reconfigure, it probably
makes sense to add a "Cache NXDOMAIN responses that do not have an SOA?"
option that defaults to "No.", along with documentation that this is a
workaround for busted upstream server configurations.

[0] <https://www.pivotaltracker.com/n/projects/956238/stories/171082087>
[1] <https://github.com/cloudfoundry/bosh-dns-release/commit/7626ca8f7ad1f300b04201df2fef11635278f41e>
[2] <https://bosh.io/docs/dns/#additional-handlers>
[3] <https://bosh.io/jobs/bosh-dns?source=github.com/cloudfoundry/bosh-dns-release&version=1.36.6#p%3dhandlers>
[4] <https://bosh.io/jobs/bosh-dns?source=github.com/cloudfoundry/bosh-dns-release&version=1.36.6#p%3dhandlers_files_glob>