Add systemd-resolved support
systemd-resolved runs its own dns server which fans out queries to other servers it knows about.
There are some global configs, but the primary discovery mechanism is via the DNS config for each
network interface.

Each network interface has it's own DNS servers and systemd-resolved tracks a "current" one that
it will send queries to. This makes it impossible to co-locate the bosh-dns configuration on an
existing interface as sometimes it would not be the "current" server and would not receive queries.

For Jammy and prior stemcells we create an alias address on the loopback interface. With systemd-resolved
we instead create a virtual network interface and bind the IP address to that. We also configure the "Domains"
for the interface to be the bosh-dns domains and any discovered bosh-dns alias doamins. This is to prevent
queries for bosh-dns addresses from being fanned out to the other dns servers.

Since systemd-resolved is already fanning out the queries to all servers in parallel, a "disable_recursors"
configuration property was added to disable recursing behavior in bosh-dns. This required some reordering of
main.go.

The "aliased_address" was also removed from the bosh-dns job as the only spot it was used didn't seem
to make sense.

Signed-off-by: Joseph Palermo <joseph.palermo@broadcom.com>