Increases timeout of flaky test
This commit increases the timeout of a known-flaky test from its default
(one second) to five seconds. As written, there's no better way to
"resolve" this issue than to wait longer. Hopefully this is long enough.

Hopefully resolves the following in CI (which is the most-common recent
failure for the 'test-unit-windows' Job)

```
  [FAILED] in [It] - C:/var/vcap/data/worker/work/containers/00003g9ti7d/tmp/build/860c4145/bosh-dns-release/src/bosh-dns/dns/main_test.go:1124 @ 07/29/24 21:59:42.589
  << Timeline

  [FAILED] Timed out after 1.002s.
  Expected
      <string>: 127.0.0.1
  to equal
      <string>: 127.0.0.3
  In [It] at: C:/var/vcap/data/worker/work/containers/00003g9ti7d/tmp/build/860c4145/bosh-dns-release/src/bosh-dns/dns/main_test.go:1124 @ 07/29/24 21:59:42.589
------------------------------
```