Fix unit test failures on dependency bump
The latest version of coredns deep copies the response object before
mutating it, which results in some slight discrepancies between our mock
object and the value stored in the cache. In the past, the actual mock
object would get mutated, so it would pass Equals testing in the units.
That no longer happens, so we need to update the mock object to
match the values we expect to actually get cached.

This fix should address the ci failure
[here](https://ci-demo.bosh-ecosystem.cf-app.com/teams/main/pipelines/bosh-dns-release/jobs/bump-deps/builds/290)